feat: migrating to app router, changes on home and fixes
