+++ src/app/projects/page.tsx
@@ -1,5 +1,6 @@
 /* eslint-disable @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-floating-promises */
-import { Page } from "@/template";
+"use client";
+
 import { useEffect, useState } from "react";
 import axios from "axios"; // Importando axios
 import { FaStar, FaCodeBranch, FaExternalLinkAlt } from "react-icons/fa";
@@ -94,89 +95,87 @@ export default function Projects() {
 	}, []);
 
 	return (
-		<Page>
-			<div className="grid grid-cols-1 lg:grid-cols-2 xl:grid-cols-3 gap-6 lg:w-5/6 w-full pt-8 lg:pt-0">
-				{loading ? (
-					<div className="flex justify-center items-center h-64 w-full">
-						<motion.div
-							className="flex space-x-2"
-							initial={{ opacity: 0 }}
-							animate={{ opacity: 1 }}
-							transition={{ repeat: Infinity, duration: 1 }}
-						>
-							{[...Array(3)].map((_, i) => (
-								<motion.span
-									key={i}
-									className="w-4 h-4 bg-white rounded-full"
-									animate={{
-										y: [0, -10, 0],
-									}}
-									transition={{
-										duration: 0.6,
-										repeat: Infinity,
-										delay: i * 0.2,
-										ease: "easeInOut",
-									}}
-								/>
-							))}
-						</motion.div>
-					</div>
-				) : (
-					repos.map((repo) => (
-						<a
-							key={repo.id}
-							href={repo.html_url}
-							target="_blank"
-							rel="noopener noreferrer"
-							className="p-6 bg-black/80 text-white rounded-lg shadow-lg transform transition duration-300 hover:scale-105 flex flex-col justify-between"
-						>
-							{repo.language && (
-								<div
-									className={`absolute top-2 left-2 px-3 py-1 text-sm font-semibold text-white rounded-lg ${
-										languageColors[repo.language] || "bg-gray-500"
-									}`}
-									style={{ marginTop: "0.5rem", marginLeft: "0.5rem" }}
-								>
-									{repo.language}
-								</div>
-							)}
-
-							<div className="flex justify-between items-center mt-8">
-								<h3 className="text-xl font-bold">{repo.name}</h3>
-								<FaExternalLinkAlt />
+		<div className="grid grid-cols-1 lg:grid-cols-2 xl:grid-cols-3 gap-6 lg:w-5/6 w-full pt-8 lg:pt-0">
+			{loading ? (
+				<div className="flex justify-center items-center h-64 w-full">
+					<motion.div
+						className="flex space-x-2"
+						initial={{ opacity: 0 }}
+						animate={{ opacity: 1 }}
+						transition={{ repeat: Infinity, duration: 1 }}
+					>
+						{[...Array(3)].map((_, i) => (
+							<motion.span
+								key={i}
+								className="w-4 h-4 bg-white rounded-full"
+								animate={{
+									y: [0, -10, 0],
+								}}
+								transition={{
+									duration: 0.6,
+									repeat: Infinity,
+									delay: i * 0.2,
+									ease: "easeInOut",
+								}}
+							/>
+						))}
+					</motion.div>
+				</div>
+			) : (
+				repos.map((repo) => (
+					<a
+						key={repo.id}
+						href={repo.html_url}
+						target="_blank"
+						rel="noopener noreferrer"
+						className="p-6 bg-black/80 text-white rounded-lg shadow-lg transform transition duration-300 hover:scale-105 flex flex-col justify-between"
+					>
+						{repo.language && (
+							<div
+								className={`absolute top-2 left-2 px-3 py-1 text-sm font-semibold text-white rounded-lg ${
+									languageColors[repo.language] || "bg-gray-500"
+								}`}
+								style={{ marginTop: "0.5rem", marginLeft: "0.5rem" }}
+							>
+								{repo.language}
 							</div>
+						)}
 
-							<p className="mt-2 text-gray-400 line-clamp-2">
-								{repo.description || "No description provided."}
-							</p>
+						<div className="flex justify-between items-center mt-8">
+							<h3 className="text-xl font-bold">{repo.name}</h3>
+							<FaExternalLinkAlt />
+						</div>
 
-							{repo.topics.length > 0 && (
-								<div className="flex flex-wrap gap-2 mt-4 line-clamp-2">
-									{repo.topics.slice(0, 6).map((topic, index) => (
-										<span
-											key={topic}
-											className={`px-3 py-1 text-sm font-semibold text-white rounded-lg ${
-												gradients[index % gradients.length]
-											}`}
-										>
-											{topic}
-										</span>
-									))}
-								</div>
-							)}
+						<p className="mt-2 text-gray-400 line-clamp-2">
+							{repo.description || "No description provided."}
+						</p>
 
-							<div className="flex justify-between mt-4 text-gray-300">
-								<span className="flex items-center">
-									<FaStar className="mr-1" /> {repo.stargazers_count}
-								</span>
-								<span className="flex items-center">
-									<FaCodeBranch className="mr-1" /> {repo.forks_count}
-								</span>
+						{repo.topics.length > 0 && (
+							<div className="flex flex-wrap gap-2 mt-4 line-clamp-2">
+								{repo.topics.slice(0, 6).map((topic, index) => (
+									<span
+										key={topic}
+										className={`px-3 py-1 text-sm font-semibold text-white rounded-lg ${
+											gradients[index % gradients.length]
+										}`}
+									>
+										{topic}
+									</span>
+								))}
 							</div>
-						</a>
-					))
-				)}
-			</div>
-		</Page>
+						)}
+
+						<div className="flex justify-between mt-4 text-gray-300">
+							<span className="flex items-center">
+								<FaStar className="mr-1" /> {repo.stargazers_count}
+							</span>
+							<span className="flex items-center">
+								<FaCodeBranch className="mr-1" /> {repo.forks_count}
+							</span>
+						</div>
+					</a>
+				))
+			)}
+		</div>
 	);
 }
