+++ src/app/tecnologies/components/TecnologyCard.tsx
