+++ src/components/ui/index.ts
@@ -1,0 +1,3 @@
+export { Footer } from "./Footer";
+export { Header } from "./Header";
+export { Sidemenu } from "./Sidemenu";
