+++ src/data/index.ts
@@ -1,1 +1,3 @@
 export { socialMedias } from "./socialMedias";
+export { experiences } from "./jobExperiences";
+export { technologies } from "./tecnologies";
