+++ src/data/tecnologies.ts
@@ -1,0 +1,135 @@
+export interface Technology {
+	technology: string;
+	years: string;
+	experienceLevel: number;
+	url: string;
+}
+
+export const technologies: Technology[] = [
+	{
+		technology: "React",
+		years: "3 y",
+		experienceLevel: 100,
+		url: "https://reactjs.org/",
+	},
+	{
+		technology: "JavaScript (ES6+)",
+		years: "5 y",
+		experienceLevel: 100,
+		url: "https://developer.mozilla.org/en-US/docs/Web/JavaScript",
+	},
+	{
+		technology: "TypeScript",
+		years: "3 y",
+		experienceLevel: 70,
+		url: "https://www.typescriptlang.org/",
+	},
+	{
+		technology: "Next.js",
+		years: "2 y",
+		experienceLevel: 70,
+		url: "https://nextjs.org/",
+	},
+	{
+		technology: "Node.js",
+		years: "2 y",
+		experienceLevel: 50,
+		url: "https://nodejs.org/",
+	},
+	{
+		technology: "Express.js",
+		years: "1 y",
+		experienceLevel: 40,
+		url: "https://expressjs.com/",
+	},
+	{
+		technology: "React Native",
+		years: "1 y",
+		experienceLevel: 40,
+		url: "https://reactnative.dev/",
+	},
+	{
+		technology: "HTML5",
+		years: "5 y",
+		experienceLevel: 100,
+		url: "https://developer.mozilla.org/en-US/docs/Web/HTML",
+	},
+	{
+		technology: "CSS3",
+		years: "5 y",
+		experienceLevel: 85,
+		url: "https://developer.mozilla.org/en-US/docs/Web/CSS",
+	},
+	{
+		technology: "SCSS/SASS",
+		years: "2 y",
+		experienceLevel: 60,
+		url: "https://sass-lang.com/",
+	},
+	{
+		technology: "Tailwind CSS",
+		years: "2 y",
+		experienceLevel: 70,
+		url: "https://tailwindcss.com/",
+	},
+	{
+		technology: "Bootstrap",
+		years: "2 y",
+		experienceLevel: 60,
+		url: "https://getbootstrap.com/",
+	},
+	{
+		technology: "Git/GitHub",
+		years: "5 y",
+		experienceLevel: 80,
+		url: "https://git-scm.com/",
+	},
+	{
+		technology: "API REST",
+		years: "3 y",
+		experienceLevel: 75,
+		url: "https://restfulapi.net/",
+	},
+	{
+		technology: "GraphQL",
+		years: "0 y",
+		experienceLevel: 30,
+		url: "https://graphql.org/",
+	},
+	{
+		technology: "Firebase",
+		years: "0 y",
+		experienceLevel: 10,
+		url: "https://firebase.google.com/",
+	},
+	{
+		technology: "SQL (MySQL/PostgreSQL)",
+		years: "0 y",
+		experienceLevel: 20,
+		url: "https://www.mysql.com/",
+	},
+	{
+		technology: "NoSQL (MongoDB)",
+		years: "0 y",
+		experienceLevel: 10,
+		url: "https://www.mongodb.com/",
+	},
+	{
+		technology: "Framer Motion",
+		years: "0 y",
+		experienceLevel: 30,
+		url: "https://www.framer.com/motion/",
+	},
+	{
+		technology: "Jest & Testing Library",
+		years: "2 y",
+		experienceLevel: 40,
+		url: "https://jestjs.io/",
+	},
+	{
+		technology: "Vite",
+		years: "2 y",
+		experienceLevel: 50,
+		url: "https://vitejs.dev/",
+	},
+];
